requirements.txt
attrs==19.2.0
coverage==4.5.2
flake8==3.7.5
Flask==2.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-HTTPAuth==3.3.0
ipaddr==2.2.0
jsonschema==2.6.0
pytest-cov==2.6.1
pytest-flake8==1.0.7
requests==2.21.0
Werkzeug==2.1.2
jinja2==3.1.2
psycopg2==2.9.3
gunicorn==20.1.0
SQLAlchemy-serializer==1.4.1